rag_pipeline/requirements.txt
requests
torch
transformers
pdfminer.six
reportlab
azure-storage-blob
openai
future
